[orga] fix "edit" button in speaker detail view
The edit button references the ID of the SpeakerProfile object, instead
of the speaker id. If these two IDs are - by chance - in sync, it works
out just fine. If there are not, pressing the edit button on one
speaker, will let you edit the details of a totally different speaker.