[common] Behave sensibly upon missing permissions
If a page is in the cfp namespace, it's probably a speaker accessing
their data, so we want to direct them to log in if they are anonymous.
If the page is in the agenda namespace, we'll probably want no login
page because participants don't have accounts.

(We should add a parameter to the mixin at some point though, so that
it's usable for plugins, too.)