Pre-fill SubmissionInfo form of CfP with track and submission_type
If the URL query string for the submission info form or the CfP landing
page contains a track or submission_types key, their value will be used
to pre-fill the submission info form. If the values do not match any in
the database, the fields will have their default as initial value.

The values consist of the ID of the track/submission_type and their
slugified name. Creation of this ID based slug is provided by the new
mixin IdBasedSlug. The original names of tracks/submission types are not
suitable because they are I18nCharFields which are not designed to be
queried.