[agenda] Make printing on Firefox slightly less horrible
Apparently the issue is that we position talk elements in the schedule
via a `top: ...px` offset from the top of that day. But Firefox, in
print mode, calculates this offset from the top of the page, so that
you'll always only get one page per day of output, and then as many
empty pages as there should be filled pages. This is stupid and nothing
we can trivially work around.

I'll recommending printing with Chromium or Vivaldi or any other browser
that works. This commit adds the fix that allows display of one page per
day in Firefox, but it's still pretty unusable due to weird page breaks
and empty pages.

Ref #625