Rewrite HTML export
Replace the django-bakery HTML export with a party home-grown
version. Internally this uses djangos test client to make "real"
WSGI requests. This makes it a little slower, but more closer to
actual requests.

This gets rid of all the custom views needed to support bakery.
The total code involved also shrinks by a lot and it easier to
understand. Additionally, the exports are now dramatically smaller,
because we only include the static and media files actually refered
to, instead of the whole static and media directories.

The new export happens in multiple stages:

1. collect all exportable page URLs (e.g. schedule, talks, ...)
2. download using `django.test.Client` and save HTML
3. extract and download CSS, JS and image files linked to in HTML
4. extract `url()`s from CSS, download the referenced files (mostly
   fonts)

Since we now follow the normal request path instead of rendering
views directly, two problems arised:

1. Get access to non-public events:

   We create an export user with access to the event inside a
   transaction. After the export is done, the transaction is rolled
   back. During development we also contemplated making the even
   public inside the transaction or using `is_html_export` for
   access control as alternative solutions.

2. Hiding elements like login or interactive search:

   Templates use the `is_html_export` variable to hide interactive
   elements. Instead of adding it to the render context directly,
   we now add it to request.META and modify the context in a
   context processor.

Other than those internal changes the new export works as a drop-in
replacement to both users and almost all of the already existing code,
including tests.