Don't show public information warning if schedule is hidden
This is a bit of a tricky decision: Some organisers use fields
differently and do not use pretalx to publish their schedules, so the
(GDPR/privacy/well-intentioned) warning that fields will be shown
publicly is actually incorrect.

At the same time, the "show schedule" setting is entirely reversible, so
"malicious" organisers could use it to avoid showing this warning.
Arguably, that's on the organiser, and since a proposal process already
usually targets public visibility, this may be an acceptable compromise.