Use dropdowns in CfP config table
This change makes it slightly harder to see at a glance what is
happening in the big CfP entry table, but in return, it makes
misconfigurations with contradicting request/require logic impossible.

This also made me come up with a way to change hierarkey forms to make
the terrible `sfom` pattern obsolete. Will be discussed in
https://github.com/raphaelm/django-hierarkey/pull/16

Thanks for this idea go to @wiffbi, who contributed it to pretix in
https://github.com/pretix/pretix/pull/1880