Remove docformatter due to upgrade instability
docformatter changes default line length and behaviour outside of
docstrings in a patch release. I'm probably too used to the utter
commitment to stability that is black, but keeping docformatter around
is not worth the maintenance effort like this. Will revert if they
commit to stronger stability guarantees. Ref PyCQA/docformatter#139