Fix language code casing
Incorrect casing of language codes meant that i18nfield always fell back
to using the language family code, disregarding the variant (if it was
upper-cased, as in Chinese, Japanese, and Portuguese). If two variants
of the same language were active at the same time in an event (only
possible with Chinese or Portuguese or German), that meant that pretalx
would not switch correctly between the two languages when rendering
user-defined content, such as on the CfP pages.